Move Parser into Separate File.
